Public API

FastHenryHelper Public API

Index

FastHenryHelper.Circle

FastHenryHelper.Comment

FastHenryHelper.Default

FastHenryHelper.Element

FastHenryHelper.End

FastHenryHelper.Equiv

FastHenryHelper.External

FastHenryHelper.Freq

FastHenryHelper.Group

FastHenryHelper.Hole

FastHenryHelper.Node

FastHenryHelper.ParseFastHenryMatResult

FastHenryHelper.Point

FastHenryHelper.Rect

FastHenryHelper.Segment

FastHenryHelper.SegmentParameters

FastHenryHelper.UniformPlane

FastHenryHelper.Units

FastHenryHelper.coilcraft1010vsgroup

FastHenryHelper.connectnodes

FastHenryHelper.elements

FastHenryHelper.helixnodes

FastHenryHelper.parsefasthenrymat

FastHenryHelper.planeconnect

FastHenryHelper.rectangulararray

FastHenryHelper.rx

FastHenryHelper.ry

FastHenryHelper.rz

FastHenryHelper.scalexyz

FastHenryHelper.terms

FastHenryHelper.transform

FastHenryHelper.txyz

FastHenryHelper.viagroup

FastHenryHelper.xyz

Elements

FastHenryHelper.Element — Type.

Subtypes of Element show FastHenry commands.

Geometric transformations can be preformed on elements with transform. Elements which require a name will automatically generate unique names if no name is provided. Groups of elements are elements.

source

FastHenryHelper.transform — Function.

transform(element, transformation_matrix)
transform!(element, transformation_matrix)

Transform (rotate, translate, scale, etc...) a element by 4x4 transform matrix tm.

Transform will modify the coordinates of Elements and the wx, wy, and wz parameters of Segment. transform of a Segment will not modify its Nodes. transform a Group containing the Nodes and Segment instead. Typically transform would only be applied to Group objects.

transform(element, transformation_matrix_list)

Return a Group of copies of element transformed by each matrix in transformation_matrix_list.

If the element is a Group, the terms of the returned group has the same keys as element, but the values are arrays of Nodes of all the copies of elements. In other words, the elements are electricaly in parallel.

source

FastHenryHelper.rx — Function.

rx(α) = [1      0       0     0;
         0    cos(α)  sin(α)  0;
         0   -sin(α)  cos(α)  0;
         0      0       0     1]

Rotation matrix for angle α around x axis.

source

FastHenryHelper.ry — Function.

ry(β) = [cos(β)   0    -sin(β)  0;
           0      1       0     0;
         sin(β)   0     cos(β)  0;
           0      0       0     1]

Rotation matrix for angle β around y axis.

source

FastHenryHelper.rz — Function.

rz(γ) = [cos(γ) sin(γ) 0  0;
        -sin(γ) cos(γ) 0  0;
          0       0    1  0;
          0       0    0  1]

Rotation matrix for angle γ around z axis.

source

FastHenryHelper.txyz — Function.

txyz(p,q,r) = [1.0 0.0 0.0  p;
               0.0 1.0 0.0  q;
               0.0 0.0 1.0  r;
               0.0 0.0 0.0 1.0]

Translation matrix. Translate a coordinate by adding [p,q,r]

source

FastHenryHelper.scalexyz — Function.

scalexyz(p,q,r) = [ p  0.0 0.0 0.0;
                   0.0  q  0.0 0.0;
                   0.0 0.0  r  0.0;
                   0.0 0.0 0.0 1.0]

Scale matrix. Scale a coordinate [x, y, z] by [p, q, r].

source

FastHenryHelper.Node — Type.

Node([name],x,y,z)
Node(xyz::Array{Float64,1})

Node objects show a FastHenry node command.

source

FastHenryHelper.xyz — Function.

xyz(n::Node)

Return the coordinate of a node.

source

FastHenryHelper.Segment — Type.

Segment([name], n1::Node, n2::Node, [parameters::SegmentParameters])
Segment([name], n1::Node, n2::Node, <keyword arguments>)

Segment objects show a FastHenry segment command.

Keyword arguments are the same as for SegmentParameters.

note: When rotating segments, the vector [wx, wy, wz] will be rotated. Default values will be used if not specified. wx, wy, wz will show after transform[!] is called on a Segment.

source

FastHenryHelper.SegmentParameters — Type.

SegmentParameters(<keyword arguments>)
SegmentParameters(parameters::SegmentParameters, <keyword arguments>)

Object to hold parameters for Segment and Default

Keyword Arguments

w – segment width

h – segment height

sigma – conductivity

rho – resistivity

wx, wy, wz – segment orientation. vector pointing along width of segment's cross section.

nhinc, nwinc – integer number of filaments in height and width

rh, rw – ratio in the height and width

The second form replaces values in parameters with any keyword parameters present.

source

FastHenryHelper.Default — Type.

Default(<keyword arguments>)
Default(parameters::SegmentParameters)

Default objects show a .default FastHenry command.

Keyword Arguments

w – segment width

h – segment height

sigma – conductivity

rho – resistivity

nhinc, nwinc – integer number of filaments in height and width

rh, rw – ratio in the height and width

source

FastHenryHelper.External — Type.

External(n1::Node, n2::Node, [portname::AbstractString])

External objects show a FastHenry .external command.

source

FastHenryHelper.End — Type.

End()

End objects show a FastHenry .end command.

source

FastHenryHelper.Equiv — Type.

Equiv(nodes::Array{Node,1})

Equiv objects show a FastHenry .equiv command

source

FastHenryHelper.Freq — Type.

Freq(min, max, [ndec])
Freq(<keyword arguments>)

Freq objects show a FastHenry .freq command.

Keyword arguments are min, max, ndec.

source

FastHenryHelper.Group — Type.

Group([elements [, terms[, units]]])
Group(<keyword arguments>)

Group objects show FastHenry commands of their elements.

Keyword Arguments

elements – Array of Element's

terms – Dict{Symbol,Node} # connections to the Group

units – Units used for the Group

Groups may be nested. Automatic name generation will create a name for each Element in the Group and all subgroups as needed when show is called. Generated names start with an underscore. Do not use these names for user named elements.

getindex, setindex!, and merge! operate on terms. push!, pop!, pushfirst!, popfirst!, append!, prepend! operate on elements.

source

FastHenryHelper.elements — Function.

elements(g::Group)
elements!(g::Group, e::Element)

Get and set elements field in a Group.

source

FastHenryHelper.terms — Function.

terms(g::Group)
terms!(g::Group, p::Dict{Symbol,Node})

Get and set the terms field in a Group.

source

FastHenryHelper.Comment — Type.

Comment(text::AbstractString)

Comment objects show a FastHenry comment.

source

FastHenryHelper.Units — Type.

Units(unitname::String)

Units objects show a FastHenry .units command.

source

FastHenryHelper.UniformPlane — Type.

UniformPlane(<keyword arguments>)

UniformPlane objects show a FastHenry uniform discretized plane command.

Keyword Arguments

name – plane name

x1,y1,z1 – coordinate of first corner

x2,y2,z2 – coordinate of second corner

x3,y3,z3 – coordinate of third corner

thick – segment thickness

seg1,seg2 – number of segments along each side of plane

segwid1,segwid2 – segment width

sigma,rho – specify conductivity or resistivity

nhinc – integer number of filaments

rh – filament ratio

relx,rely,relz– see FastHenry documentation

nodes – Array of Node connection points to plane

holes – Array of Hole (Point, Rect, or Circle)

source

FastHenryHelper.Hole — Type.

Subtypes of Hole are holes in a plane.

Subtypes

Point

Rect

Circle

source

FastHenryHelper.Point — Type.

Point(x, y, z)
Point(<keyword arguments>)

Point objects are used for point holes in a UniformPlane.

Keyword arguments are x, y, z.

source

FastHenryHelper.Rect — Type.

Rect(x1, y1, z1, x2, y2, z2)
Rect(<keyword arguments>)

Rect objects are used for rectangular holes in a UniformPlane.

Keyword arguments are x1, y1, z1, x2, y2, z2.

source

FastHenryHelper.Circle — Type.

Circle(x, y, z, r)
Circle(<keyword arguments>)

Circle objects are used for circular holes in a UniformPlane.

Keyword arguments are x, y, z, r.

source

Groups and Arrays

FastHenryHelper.connectnodes — Function.

connectnodes(nodes::Array{Node,1}, [parameters::SegmentParameters])

Returns an array of Segments connecting Nodes.

source

FastHenryHelper.helixnodes — Function.

    helixnodes(radius, pitch, radians, [radiansperpoint=π/4])

Returns an array of Nodes in a helix.

source

FastHenryHelper.viagroup — Function.

viagroup(radius, height, n, [parameters::SegmentParameters [,topequiv [,botequiv]]])
viagroup(<keyword arguments>)

Returns a Group of Segments for the barrel of a via.

Keyword Arguments

radius – radius of via

height – height of via

h – thickness of copper plating

n – number of segment around via

sigma,rho – specify conductivity or resistivity

nhinc – number of filaments per segment

rh – ratio of filament width

topequiv – include .equiv for all top nodes. default = true

botequiv – include .equiv for all bot nodes. default = true

Terminals are in the center labeled :top and :bot. Only SegmentParameters which match keyword arguments are used. Specify n=0 to bypass via (.equiv top bot). Terminals :alltop and :allbot are arrays of all top and bottom nodes. :alltop, :allbot, topequiv, and botequiv are intended for connections to planes.

source

FastHenryHelper.coilcraft1010vsgroup — Function.

coilcraft1010vsgroup(partnumber, <keyword arguments>)

Returns a Group for a Coilcraft 1010VS series inductor.

Part Numbers

1010VS-23NME

1010VS-46NME

1010VS-79NME

1010VS-111ME

1010VS-141ME

Keyword Arguments

nhinc, nwinc – integer number of filaments in height and width

rh, rw – ratio in the height and width

Terminals are :a and :b.

source

Utility

FastHenryHelper.rectangulararray — Function.

rectangulararray(x[,y[,z]])
rectangulararray(<keyword arguments>)

Used with transform to create rectangular arrays of elements.

rectangulararray returns a iterable object which returns a sequence of transformation matrices, which when passed to transform will produce a rectangular array of elements. x, y, and z are iterable objects which specify the offsets along the x, y, and z axis. Unspecified arguments are zero.

source

FastHenryHelper.planeconnect — Function.

planeconnect(nodes::Array{Node,1})
planeconnect(node::Node)

Provides objects needed to connect to a plane.

Returns a tuple of a deepcopy of the nodes passed and a group of Equiv objects connecting the nodes passed to the nodes returned.

(planenodearray, equivgroup) = planeconnect(nodearray)

source

FastHenryHelper.parsefasthenrymat — Function.

parsefasthenrymat(io::IO)
parsefasthenrymat(filename::AbstractString)

parses the .mat output file from FastHenry

returns type ParseFastHenryResult

source

FastHenryHelper.ParseFastHenryMatResult — Type.

ParseFastHenryMatResult

Result of parsing FastHenry .mat file.

Fields

portnames – array of port names, index is row number in impedance matrix

frequencies – frequencies at which impedance matrix is computed

impedance – impedance matrix at each frequency. impedance[row, col, frequency]

source